Handle delayed snapshot ID updates better
Since Spotify doesn't immediately update the snapshot ID of a playlist
on change (anymore??), we need to check the track count. Due to how
the FullPlaylist struct is implemented, there was a bug when caching
FullPlaylists that caused the Tracks field to be empty. This is fixed
now, somewhat hackily, but fixed.